ContinuousNormalizingFlows

Documentation for ContinuousNormalizingFlows.

ContinuousNormalizingFlows.ContinuousNormalizingFlows

ContinuousNormalizingFlows.CondFFJORD

ContinuousNormalizingFlows.CondPlanar

ContinuousNormalizingFlows.CondRNODE

ContinuousNormalizingFlows.FFJORD

ContinuousNormalizingFlows.Planar

ContinuousNormalizingFlows.PlanarLayer

ContinuousNormalizingFlows.RNODE

ContinuousNormalizingFlows.ContinuousNormalizingFlows — Module

Implementations of Infinitesimal Continuous Normalizing Flows Algorithms in Julia

ContinuousNormalizingFlows.CondFFJORD — Type

Implementation of FFJORD (Conditional Version)

ContinuousNormalizingFlows.CondPlanar — Type

Implementation of Planar (Conditional Version)

ContinuousNormalizingFlows.CondRNODE — Type

Implementation of RNODE (Conditional Version)

ContinuousNormalizingFlows.FFJORD — Type

Implementation of FFJORD from

Grathwohl, Will, Ricky TQ Chen, Jesse Bettencourt, Ilya Sutskever, and David Duvenaud. "Ffjord: Free-form continuous dynamics for scalable reversible generative models." arXiv preprint arXiv:1810.01367 (2018).

ContinuousNormalizingFlows.Planar — Type

Implementation of Planar Flows from

Chen, Ricky TQ, Yulia Rubanova, Jesse Bettencourt, and David Duvenaud. "Neural Ordinary Differential Equations." arXiv preprint arXiv:1806.07366 (2018).

ContinuousNormalizingFlows.PlanarLayer — Type

Implementation of Planar Layer from

Chen, Ricky TQ, Yulia Rubanova, Jesse Bettencourt, and David Duvenaud. "Neural Ordinary Differential Equations." arXiv preprint arXiv:1806.07366 (2018).

ContinuousNormalizingFlows.RNODE — Type

Implementation of RNODE from

Finlay, Chris, Jörn-Henrik Jacobsen, Levon Nurbekyan, and Adam M. Oberman. "How to train your neural ODE: the world of Jacobian and kinetic regularization." arXiv preprint arXiv:2002.02798 (2020).